% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fetch_school_type.R
\name{fetch_school_type}
\alias{fetch_school_type}
\title{Fetch school types for BNs}
\usage{
fetch_school_type(bns, sqr_official_dir, report_type = FALSE,
  collapse_types = TRUE)
}
\arguments{
\item{bns}{A vector or list of BNs}

\item{sqr_official_dir}{The directory of SAS datasets to use. Must be an SQR official directory.}

\item{report_type}{Report report_type instead of school_type? Default: FALSE}

\item{collapse_types}{Should multiple school_types be collapsed when possible? Default: TRUE}
}
\value{
A vector of school_types or report_types
}
\description{
Fetch a vector of school types from a vector of BNs.
}
\examples{
my.data$school_type <- fetch_school_type(my.data$bn, sqr_2014_local)
}

