% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_sqr_dataset.R
\name{load_sqr_dataset}
\alias{load_sqr_dataset}
\title{Import a SQR/PR SAS dataset}
\usage{
load_sqr_dataset(dataset_name, sas_dir)
}
\arguments{
\item{dataset_name}{The name of the SAS dataset you want to import.}

\item{sas_dir}{The directory of the dataset you want to import. Use helper functions if it is a common directory.}
}
\value{
A dataframe of the imported SAS dataset.
}
\description{
Import a SQR/PR SAS dataset from a given directory.
}
\examples{
demmutableforstu <- load_sqr_dataset('demmutableforstu', sqr_ems_unofficial_2014())
}

