% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pr_hs_official_2013.R
\name{pr_hs_official_2013}
\alias{pr_hs_official_2013}
\title{Get the directory of the PR official HS SAS datasets for 2013}
\usage{
pr_hs_official_2013()
}
\value{
The directory.
}
\description{
Return the directory of the PR Official HS SAS datasets
}
\examples{
pr_dir <- pr_hs_official_2013()
}

