% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query_sql_server.R
\name{query_sql_server}
\alias{query_sql_server}
\title{Run a query on a SQL server.}
\usage{
query_sql_server(dbh, query)
}
\arguments{
\item{dbh}{A database handle. Use sql15(), sql27(), or sqlord() for typical DOE SQL servers.}

\item{query}{The query you want to run. Make sure to include the database and table schema in the table name (e.g. \code{spr_int.prl.extract_school_data}).}
}
\value{
A dataframe of the data returned by the query.
}
\description{
Run a query on a sQL server and return the result as a dataframe.
}
\examples{
school_data <- query_sql_server(sql27(), "select dbn, quality_review_score from spr_int.prl.extract_school_data where year = 2013")
}

