% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql15.R
\name{sql15}
\alias{sql15}
\title{Get a database handle for SQL Server 15.}
\usage{
sql15()
}
\value{
A database handle.
}
\description{
Get a database handle for mtsqlvs15/mtsqlins15.
}
\examples{
dbh15 <- sql15()
}

