% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql27.R
\name{sql27}
\alias{sql27}
\title{Get a database handle for SQL Server 27.}
\usage{
sql27()
}
\value{
A database handle.
}
\description{
Get a database handle for mtsqlvs27/mtsqlins27.
}
\examples{
dbh27 <- sql27()
}

