% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sqlord.R
\name{sqlord}
\alias{sqlord}
\title{Get a database handle for ORD's SQL server.}
\usage{
sqlord()
}
\value{
A database handle.
}
\description{
Get a database handle for ORD's SQL server.
}
\examples{
dbhord <- sqlord()
}

