% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sqr_hs_unofficial_2014.R
\name{sqr_hs_unofficial_2014}
\alias{sqr_hs_unofficial_2014}
\title{Get the directory of the SQR Unofficial HS SAS datasets for 2014}
\usage{
sqr_hs_unofficial_2014()
}
\value{
The directory.
}
\description{
Return the directory of the SQR Unofficial HS SAS datasets.
}
\examples{
sqr_dir <- sqr_hs_unofficial_2014()
}

