% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sqr_official_2014.R
\name{sqr_official_2014}
\alias{sqr_official_2014}
\title{Get the directory of the SQR official SAS datasets for 2014 (for schall*_targets only)}
\usage{
sqr_official_2014()
}
\value{
The directory.
}
\description{
Return the directory of the SQR Official SAS datasets (schall*_targets only)
}
\examples{
sqr_dir <- sqr_official_2014()
}

